requirements.txt
Django
djangorestframework
elasticsearch
python-dotenv